# backend/openapi.yaml
openapi: 3.0.3
info:
  title: progctf — The Voyage API
  version: 0.1.0
  description: |
    Pirate-themed CTF backend (Bosun). Player REST API + admin API.
    JWT auth via httpOnly cookie (`progctf_session` for player, `progctf_admin` for admin)
    or `Authorization: Bearer <token>` header.
servers:
  - url: http://localhost:4000
    description: Local dev
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: progctf_session
    adminCookie:
      type: apiKey
      in: cookie
      name: progctf_admin
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    OkResponse:
      type: object
      properties:
        ok: { type: boolean, example: true }
        message: { type: string }
    ErrorResponse:
      type: object
      required: [ok, code, message]
      properties:
        ok: { type: boolean, example: false }
        code:
          type: string
          enum:
            - ERR_BAD_INPUT
            - ERR_BAD_FORMAT
            - ERR_BAD_CREDENTIALS
            - ERR_PIRATE_EXISTS
            - ERR_BANNED
            - ERR_NO_AUTH
            - ERR_BAD_AUTH
            - ERR_NOT_ADMIN
            - ERR_NO_REFRESH
            - ERR_BAD_REFRESH
            - ERR_NO_CREW
            - ERR_ALREADY_ON_CREW
            - ERR_CREW_NAME_TAKEN
            - ERR_CREW_BANNED
            - ERR_NO_ISLAND
            - ERR_NOT_PUBLISHED
            - ERR_NO_WHISPER
            - ERR_BAD_WHISPER
            - ERR_RATE_LIMITED
            - ERR_ALREADY_PLUNDERED
            - ERR_WRONG_TREASURE
            - ERR_VOYAGE_FROZEN
            - ERR_SLUG_TAKEN
            - ERR_WHISPER_EXISTS
            - ERR_NO_PIRATE
            - ERR_NO_UPDATES
            - ERR_SANDBOX_DOWN
        message: { type: string, example: "That ain't the right Treasure, sailor." }
        retryAfter: { type: integer }
    PirateBrief:
      type: object
      properties:
        id: { type: string, format: uuid }
        handle: { type: string }
        role: { type: string, enum: [pirate, admin] }
        crew_id: { type: string, format: uuid, nullable: true }
    PirateMe:
      type: object
      required: [id, email, handle, role, crew]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        handle: { type: string }
        role: { type: string, enum: [pirate, admin] }
        crew:
          oneOf:
            - type: 'null'
            - type: object
              required: [id, name, flag_emoji]
              properties:
                id: { type: string, format: uuid }
                name: { type: string }
                flag_emoji: { type: string, nullable: true }
    CrewProfile:
      type: object
      required: [id, name, score, members, solved]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        flag_emoji: { type: string, nullable: true }
        score: { type: integer }
        rank: { type: integer, nullable: true }
        members:
          type: array
          items:
            type: object
            properties:
              handle: { type: string }
        solved:
          type: array
          items:
            type: object
            properties:
              island_slug: { type: string }
              island_title: { type: string }
              category: { type: string }
              difficulty: { type: string, enum: [port, open_sea, cursed_depths] }
              awarded_points: { type: integer }
              first_blood: { type: boolean }
              solved_at: { type: string, format: date-time }
...
paths:
...
  /api/auth/sign-articles:
    post:
      tags: [auth]
      summary: Register a Pirate (rate-limited 10/min/IP)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, handle, password]
              properties:
                email: { type: string, format: email }
                handle: { type: string }
                password: { type: string, format: password, minLength: 8 }
      responses:
        '201':
          description: Pirate created and auto-boarded (cookies set)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OkResponse'
                  - type: object
                    properties:
                      pirate: { $ref: '#/components/schemas/PirateBrief' }
        '400': { description: Bad input, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: Email or handle taken, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '429': { description: Rate limited }
  /api/auth/board:
    post:
      tags: [auth]
      summary: Login (rate-limited 10/min/IP); JWT issued in httpOnly cookie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200': { description: Boarded }
        '401': { description: Bad credentials, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: Banned }
  /api/auth/refresh:
    post:
      tags: [auth]
      summary: Rotate access + refresh tokens
      responses:
        '200': { description: ok }
        '401': { description: missing/invalid refresh }
...
  /api/auth/me:
    get:
      tags: [auth]
      summary: Identity probe — returns the calling Pirate + Crew (frontend uses this since the JWT is httpOnly)
      security: [{ sessionCookie: [] }, { bearer: [] }]
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PirateMe' }
        '401':
          description: missing or invalid auth
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '404':
          description: Pirate row no longer exists
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/auth/logout:
    post:
      tags: [auth]
      summary: Logout — clears BOTH the player cookie and the admin cookie. Themed alias of /disembark for the frontend.
      security: [{ sessionCookie: [] }, { bearer: [] }]
      responses:
        '200':
          description: cookies cleared
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OkResponse'
                  - type: object
                    properties:
                      message: { type: string, example: "Walked the plank." }
        '401':
          description: missing or invalid auth
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/crews/:
    post:
      tags: [crews]
      summary: Create a Crew
      security: [{ sessionCookie: [] }, { bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                flag_emoji: { type: string }
      responses:
        '201': { description: created }
        '409': { description: name taken / already on crew }
  /api/crews/{name}:
    get:
      tags: [crews]
      summary: Crew profile + members + solved Islands (sorted newest-first)
      security: [{ sessionCookie: [] }, { bearer: [] }]
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string, minLength: 2, maxLength: 40 }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CrewProfile' }
        '400':
          description: malformed crew name
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '404':
          description: no crew with that name
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/crews/join:
    post:
      tags: [crews]
      summary: Join a Crew via invite code
      security: [{ sessionCookie: [] }, { bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invite_code]
              properties:
                invite_code: { type: string }
      responses:
        '200': { description: joined }
        '404': { description: no such crew }
...